Partially fixed timer (Read Desc.)
The Timer, which has been jumping around seemingly randomly, is not broken due to an error in our code, but rather an a glitch in the font which we have been using. Annoyingly, the fix cannot be pushed onto Git, so here are the instructions: Find "Electronic Highway Sign SDF (TMP_Font Asset)" in "Assets/TextMesh Pro/Examples & Extras/Resources/Fonts & Materials" Then, at the bottom of the menu, expand "Character Table." The Character glyphs are screwed up, and you'll have to fix them. First, change each character so that the glyph it's supposed to be is the last number of the Unicode. Second, if you get two nines, delete the second one. Third, change the glyph ID to the following (If it is already these numbers don't change it): 0 is 19, 1 is 20, 2 is 21, 3 is 27, 4 is 25, 5 is 24, 6 is 22, 7 is 23, 8 is 26, and 9 is 28. After everything is set, make sure it saves and test, talk to me if you have any problems.